requirements.txt
requests
pysocks
gspread
oauth2client